Add codexw: profile-aware multi-pass Codex review wrapper
- Split monolithic __main__.py into 13 focused modules
- Model fallback chain (recency-biased, 5-model window)
- Reasoning effort fallback (xhigh → high → medium → low)
- State reuse across passes
- Fallback YAML parser/writer (no PyYAML required)
- 28 regression tests
- architecture.md + features-and-usecases.md documentation
