+++ .pre-commit-hooks.yaml
@@ -35,11 +35,9 @@
   stages: [manual]
   verbose: true
 
-# Nobody should ever use these hooks in production. They're just for testing PRs in
-# the duolingo/pre-commit-hooks repo more easily without having to tag and push
-# temporary images to Docker Hub. Usage: edit a consumer repo's hook config to
-# instead declare `id: duolingo-dev` or `id: sync-ai-rules-dev` and `rev: <PR branch SHA>`,
-# then run `pre-commit run <hook-id> --all-files`
+# Dev hooks for testing PRs in the duolingo/pre-commit-hooks repo.
+# Usage: edit a consumer repo's hook config to declare `id: duolingo-dev`
+# and `rev: <PR branch SHA>`, then run `pre-commit run <hook-id> --all-files`
 - id: duolingo-dev
   name: Duolingo (dev)
   entry: /entry
